Rename some folders and files to match the new topic
